refactor(codex-manager): extract inline command handlers (phase 3)
Continue the lib/codex-manager.ts decomposition (audit roadmap §4.1.1
phase 3, stacked on the registry branch, PR #535). Moves 1,218 net lines
out of codex-manager.ts (3,484 -> 2,266) by extracting the remaining
inline handler bodies and the module-scope helpers they closure-captured.
All moves are verbatim; zero behavior change. Every added line in
codex-manager.ts is an import.

Command-body extractions:
- check: runHealthCheck (~325 lines) -> lib/codex-manager/health-check.ts.
  Also reused by the login dashboard quick/deep-check actions, so it lives
  beside repair-commands.ts rather than under commands/. Imports library
  deps directly, exactly like repair-commands.ts does.
- best: parseBestArgs + printBestUsage (~60 lines) ->
  lib/codex-manager/commands/best.ts. Still injected through
  BestCommandDeps by the dispatcher, so the command interface and tests
  are untouched.
- login (data layer, ~460 lines) -> lib/codex-manager/login-menu-data.ts:
  menu quota auto-refresh targeting (countMenuQuotaRefreshTargets,
  refreshQuotaCacheForMenu, DEFAULT_MENU_QUOTA_REFRESH_TTL_MS), the
  ExistingAccountInfo row mapping + ready-first ordering
  (toExistingAccountInfo), runtime current-account resolution, and the
  Codex CLI selection drift sync.

Shared helpers relocated (were module-private, closure-captured by the
repair/forecast/report/best dependency factories and the blocks above):
- lib/codex-manager/quota-cache-helpers.ts (~145 lines):
  getPersistedQuotaViewForAccount, updateQuotaCacheForAccount,
  cloneQuotaCacheData, pruneUnsafeQuotaEmailCacheEntry,
  DEFAULT_LIVE_PROBE_MODEL.
- lib/codex-manager/account-credentials.ts (~70 lines):
  hasUsableAccessToken, hasLikelyInvalidRefreshToken,
  resolveStoredAccountIdentity, applyTokenAccountIdentity.
- lib/codex-manager/persist-selected-account.ts (~125 lines):
  persistAndSyncSelectedAccount, shared by switch/best (deps injection)
  and the login backup-restore flow.

codex-manager.ts now imports these and keeps injecting them through the
existing deps interfaces (createRepairCommandDeps, runForecast/runBest/
report wrappers), so no command-module interface changed. No new module
imports codex-manager.ts (no cycles); nothing previously exported moved.

Remaining for phase 4: the interactive login control loop and OAuth
machinery (runAuthLogin/runAuthLoginFlow, handleManageAction + manage
helpers, runOAuthFlow/runSignInFlow, the sign-in/backup prompts,
runActionPanel/waitForMenuReturn, persistAccountPool). These share
closure-mutable menu state (pending refresh promise, status string,
skip/generation counters) and label-based control flow, so extracting
them is a larger, riskier move that deserves its own PR.

Verification: tsc clean; eslint --max-warnings=0 clean on touched files;
the 64 suites matching codex-manager/runCodexMultiAuthCli show identical
results to the base branch (822 passed, 6 skipped, 3 failed - the known
Windows-path environment failures in codex-bin-wrapper.test.ts, present
on base).
